Adding better error messaging for port ID check
Problem:
There's not enough information in the InvalidFixedPortIDError to
diagnose the real error.

Solution:
Include if the ID was for a service or message type and include the full
path. This allows the user to see which ID range they violated or if
this is just a path error.

Testing:
Fed invalid identifiers and verified that the message is more
informative.