Proposal with failing tests
This is a proposal that I want to discuss before I go ahead and fix the
tests. The TaggedUnion has a tag and an UntaggedUnion but is a data structure. You'll see this gets rid of some specalized code in pydsdl and will enable more generalized code in pydsdl consumers.