Allow relative target paths (#110)
Fix for issue #109. This relaxes some logic that required existing paths
before creating a DSDLFile abstraction while adding a requirement that
ReadableDSDLFile(s) are created with files that exist (i.e. it can't be
"readable" if it doesn't exist). At the same time, we clarified (and
fixed) the difference between the `root_namespace_directories_or_names`
and `lookup_directories arguments` of read_files. Our new guidance is to
dis-use `lookup_directories` unless there is a need to separate target
lookup paths from dependent type look up paths. This was an unstated
design goal that I forgot about and subsequently mis-implemented. This
does change the behaviour of the API slightly which is why I've bumped
the minor version instead of just the patch.